Don't disconnect out-of-LOS tentacles (#11213)
In 0.21-a0-253-g6b34986 we began enforcing PROX_CLOSE_TO_PLAYER even
when the monster was placed with MG_FORCE_PLACE.  But the tentacle code
used that combination, resulting in a failure to place and a can't-
happen message once the tentacled monster left LOS radius.

Use PROX_ANYWHERE instead.